+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Breaking Changes
+
+- `calcite-modal` - `close-label` prop is now renamed to `intl-close` for consistency (#466)
+- `calcite-date` - `prevMonthLabel` and `nextMonthLabel` updated to `intlPrevMonth` and `intlNextMonth` (#97)
 - `calcite-switch` - `switched` boolean has been added to `calciteSwitchChange` event detail
 
 ## [v1.0.0-beta.31]
@@ -516,6 +520,7 @@ Fix issue with previous release.
 
 First initial beta release.
 
+[unreleased]: https://github.com/Esri/calcite-components/compare/master...v1.0.0-beta.31 "Unreleased"
 [v1.0.0-beta.31]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.30...v1.0.0-beta.31 "v1.0.0-beta.31"
 [v1.0.0-beta.30]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.29...v1.0.0-beta.30 "v1.0.0-beta.30"
 [v1.0.0-beta.29]: https://github.com/Esri/calcite-components/compare/v1.0.0-beta.28...v1.0.0-beta.29 "v1.0.0-beta.29"